Simplify API
Finally decided on a certain API style.
